# swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Catalog API",
    "description": "Microservicio de Catalog - ECommerce Architecture - Soporte multiidioma con Accept-Language header (es, en)",
    "version": "v1"
  },
  "paths": {
...
    "/v1/products": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "take",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "ids",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDtoDataCollection"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDtoDataCollection"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDtoDataCollection"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductCreateCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/v1/products/{id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/products/search": {
      "get": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "Query",
            "in": "query",
            "schema": {
              "maxLength": 200,
              "type": "string"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "maximum": 2147483647,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductSortField"
            }
          },
          {
            "name": "SortOrder",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SortOrder"
            }
          },
          {
            "name": "CategoryId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "BrandIds",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "MinPrice",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "MaxPrice",
            "in": "query",
            "schema": {
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "InStock",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsFeatured",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "HasDiscount",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "MinRating",
            "in": "query",
            "schema": {
              "maximum": 5,
              "minimum": 0,
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/products/search/advanced": {
      "post": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAdvancedSearchRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAdvancedSearchRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductAdvancedSearchRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductAdvancedSearchResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductAdvancedSearchResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductAdvancedSearchResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/v1/products/admin/clear-search-cache": {
      "post": {
        "tags": [
          "Product"
        ],
        "parameters": [
          {
            "name": "Accept-Language",
            "in": "header",
            "description": "Idioma preferido para el contenido (es = Español, en = English). Default: es",
            "schema": {
              "enum": [
                "es",
                "en"
              ],
              "type": "string",
              "default": "es"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProductCreateCommand": {
        "type": "object",
        "properties": {
          "nameSpanish": {
            "type": "string",
            "nullable": true
          },
          "nameEnglish": {
            "type": "string",
            "nullable": true
          },
          "descriptionSpanish": {
            "type": "string",
            "nullable": true
          },
          "descriptionEnglish": {
            "type": "string",
            "nullable": true
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "slug": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "originalPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "discountPercentage": {
            "type": "number",
            "format": "double"
          },
          "taxRate": {
            "type": "number",
            "format": "double"
          },
          "images": {
            "type": "string",
            "nullable": true
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "metaKeywords": {
            "type": "string",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          },
          "isFeatured": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "ProductDtoDataCollection": {
        "type": "object",
        "properties": {
          "hasItems": {
            "type": "boolean",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "ProductSearchResponse": {
        "type": "object",
        "properties": {
          "hasItems": {
            "type": "boolean",
            "readOnly": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductDto"
            },
            "nullable": true
          },
          "total": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pages": {
            "type": "integer",
            "format": "int32"
          },
          "metadata": {
            "$ref": "#/components/schemas/SearchMetadata"
          }
        },
        "additionalProperties": false
      },
      "ProductSortField": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6
        ],
        "type": "integer",
        "format": "int32"
      },
...
      "SortOrder": {
        "enum": [
          0,
          1
        ],
        "type": "integer",
        "format": "int32"
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Ingrese el token JWT en el formato: Bearer {token}",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      },
      "ApiKey": {
        "type": "apiKey",
        "description": "Ingrese el API Key para comunicación entre servicios",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ],
      "ApiKey": [ ]
    }
  ]
}